package geometricFigures renamed
