+++ Swing-Lab-Work/src/gui/AddDialog.java
@@ -3,9 +3,9 @@ package gui;
 import converter.Converter;
 import exceptions.VolumeException;
 import exceptions.OverflowException;
-import geometricFigures.Circle;
-import geometricFigures.Square;
-import geometricFigures.Triangle;
+import planeFigures.Circle;
+import planeFigures.Square;
+import planeFigures.Triangle;
 import shapes.Cube;
 import shapes.Cylinder;
 import shapes.Pyramid;
